# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: JIRA CDC Operator API
  description: |
    REST API for JIRA Change Data Capture Kubernetes Operator managed API operand.
    
    This API provides endpoints for monitoring and managing JIRA to Git synchronization
    operations, including project status, task management, and system health monitoring.
    
    The API is designed to be consumed by the React UI operand and external monitoring systems.
  version: 1.0.0
  contact:
    name: JIRA CDC Operator
    url: https://github.com/company/jira-cdc-operator
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development API operand
  - url: https://jiracdc-api.{namespace}.svc.cluster.local:8080/api/v1
    description: Kubernetes cluster API operand
paths:
  /projects:
    get:
      summary: List synchronized JIRA projects
      description: |
        Returns list of JIRA projects currently being synchronized by this operator instance.
        Projects are derived from active JiraCDC resources in the cluster.
      operationId: listProjects
      tags:
        - Projects
      responses:
        '200':
          description: List of synchronized projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSummary'
              examples:
                multiple_projects:
                  summary: Multiple projects example
                  value:
                    - projectKey: "PROJ1"
                      name: "Project Alpha"
                      status: "Current"
                      lastSyncTime: "2025-09-19T14:25:00Z"
                      syncedIssueCount: 156
                      gitRepository: "git@github.com:company/proj1-mirror.git"
                    - projectKey: "PROJ2"
                      name: "Project Beta"
                      status: "Syncing"
                      lastSyncTime: "2025-09-19T14:20:00Z"
                      syncedIssueCount: 89
                      gitRepository: "git@github.com:company/proj2-mirror.git"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectKey}:
    get:
      summary: Get project details
      description: |
        Returns detailed information about a specific synchronized JIRA project,
        including operand status, recent tasks, and synchronization configuration.
      operationId: getProject
      tags:
        - Projects
      parameters:
        - name: projectKey
          in: path
          required: true
          description: JIRA project key
          schema:
            type: string
            pattern: '^[A-Z][A-Z0-9_]*$'
            example: "PROJ"
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetails'
              examples:
                project_details:
                  summary: Project details example
                  value:
                    projectKey: "PROJ"
                    name: "Sample Project"
                    status: "Current"
                    lastSyncTime: "2025-09-19T14:25:00Z"
                    syncedIssueCount: 156
                    gitRepository: "git@github.com:company/proj-mirror.git"
                    jiraInstance:
                      baseURL: "https://company.atlassian.net"
                      proxyEnabled: false
                    syncConfig:
                      interval: "5m"
                      activeIssuesOnly: true
                      bootstrap: true
                    operandStatus:
                      api:
                        ready: true
                        replicas: 2
                        readyReplicas: 2
                        endpoint: "http://jiracdc-api.default.svc.cluster.local:8080"
                      ui:
                        ready: true
                        replicas: 1
                        readyReplicas: 1
                        endpoint: "http://jiracdc-ui.default.svc.cluster.local:3000"
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: Project not found example
                  value:
                    error: "NotFound"
                    message: "Project INVALID not found"
                    timestamp: "2025-09-19T14:30:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectKey}/sync:
    post:
      summary: Trigger manual synchronization
      description: |
        Triggers a manual synchronization operation for the specified project.
        This creates a new CDC task that will be processed asynchronously.
        
        Supported sync types:
        - `bootstrap`: Full synchronization of all matching issues
        - `reconciliation`: Incremental synchronization of updated issues
        - `maintenance`: Cleanup and consistency checks
      operationId: triggerSync
      tags:
        - Projects
      parameters:
        - name: projectKey
          in: path
          required: true
          description: JIRA project key
          schema:
            type: string
            pattern: '^[A-Z][A-Z0-9_]*$'
            example: "PROJ"
      requestBody:
        description: Sync operation parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
            examples:
              bootstrap_sync:
                summary: Bootstrap synchronization
                value:
                  type: "bootstrap"
                  forceRefresh: true
                  issueFilter: "project = PROJ"
              incremental_sync:
                summary: Incremental synchronization
                value:
                  type: "reconciliation"
                  forceRefresh: false
                  issueFilter: "status != Done AND updated >= -7d"
      responses:
        '202':
          description: Sync operation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              examples:
                task_started:
                  summary: Task started successfully
                  value:
                    taskId: "task-123e4567-e89b-12d3-a456-426614174000"
                    status: "started"
                    message: "Bootstrap task started successfully"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_type:
                  summary: Invalid sync type
                  value:
                    error: "InvalidRequest"
                    message: "Invalid sync type: invalid_type"
                    timestamp: "2025-09-19T14:30:00Z"
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks:
    get:
      summary: List CDC tasks
      description: |
        Returns list of bootstrap, reconciliation, and maintenance tasks.
        Supports filtering by status, type, and project key.
        
        Tasks are returned in reverse chronological order (newest first).
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - name: status
          in: query
          description: Filter tasks by status
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
          example: "running"
        - name: type
          in: query
          description: Filter tasks by type
          schema:
            type: string
            enum: [bootstrap, reconciliation, maintenance]
          example: "bootstrap"
        - name: projectKey
          in: query
          description: Filter tasks by project key
          schema:
            type: string
            pattern: '^[A-Z][A-Z0-9_]*$'
            example: "PROJ"
        - name: limit
          in: query
          description: Maximum number of tasks to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 20
        - name: offset
          in: query
          description: Number of tasks to skip (for pagination)
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskSummary'
              examples:
                task_list:
                  summary: Mixed task statuses
                  value:
                    - id: "task-123e4567-e89b-12d3-a456-426614174000"
                      type: "bootstrap"
                      status: "completed"
                      projectKey: "PROJ1"
                      createdAt: "2025-09-19T14:00:00Z"
                      startedAt: "2025-09-19T14:01:00Z"
                      completedAt: "2025-09-19T14:15:00Z"
                      progress:
                        totalItems: 100
                        processedItems: 100
                        percentComplete: 100.0
                    - id: "task-456e7890-e12b-34c5-d678-901234567890"
                      type: "reconciliation"
                      status: "running"
                      projectKey: "PROJ2"
                      createdAt: "2025-09-19T14:10:00Z"
                      startedAt: "2025-09-19T14:11:00Z"
                      progress:
                        totalItems: 50
                        processedItems: 25
                        percentComplete: 50.0
                        estimatedTimeRemaining: "5m30s"
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{taskId}:
    get:
      summary: Get task details
      description: |
        Returns detailed information about a specific CDC task,
        including configuration, progress, operations, and error details.
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task identifier (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetails'
              examples:
                completed_task:
                  summary: Completed bootstrap task
                  value:
                    id: "task-123e4567-e89b-12d3-a456-426614174000"
                    type: "bootstrap"
                    status: "completed"
                    projectKey: "PROJ"
                    createdAt: "2025-09-19T14:00:00Z"
                    startedAt: "2025-09-19T14:01:00Z"
                    completedAt: "2025-09-19T14:15:00Z"
                    progress:
                      totalItems: 100
                      processedItems: 100
                      percentComplete: 100.0
                    configuration:
                      issueFilter: "project = PROJ"
                      forceRefresh: true
                    createdBy: "jiracdc-operator"
                    operations:
                      - id: "op-456e7890-e12b-34c5-d678-901234567890"
                        issueKey: "PROJ-123"
                        operationType: "create"
                        status: "completed"
                        processedAt: "2025-09-19T14:05:00Z"
                        gitOperation:
                          action: "file created"
                          filePath: "PROJ-123.md"
                          commitMessage: "feat(PROJ-123): add new issue"
                          commitHash: "abc123def456"
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{taskId}/cancel:
    post:
      summary: Cancel task
      description: |
        Cancels a pending or running task. Completed, failed, or already 
        cancelled tasks cannot be cancelled.
        
        The cancellation is graceful - running operations will complete
        before the task is marked as cancelled.
      operationId: cancelTask
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task identifier (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Task cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              examples:
                task_cancelled:
                  summary: Task cancelled successfully
                  value:
                    taskId: "task-123e4567-e89b-12d3-a456-426614174000"
                    status: "cancelled"
                    message: "Task cancelled successfully"
        '400':
          description: Task cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_completed:
                  summary: Task already completed
                  value:
                    error: "InvalidRequest"
                    message: "Task with status 'completed' cannot be cancelled"
                    timestamp: "2025-09-19T14:30:00Z"
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /issues/{issueKey}:
    get:
      summary: Get issue sync status
      description: |
        Returns synchronization status and metadata for a specific JIRA issue,
        including git file information and sync history.
      operationId: getIssueStatus
      tags:
        - Issues
      parameters:
        - name: issueKey
          in: path
          required: true
          description: JIRA issue key
          schema:
            type: string
            pattern: '^[A-Z][A-Z0-9_]*-[0-9]+$'
            example: "PROJ-123"
      responses:
        '200':
          description: Issue sync status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueSyncStatus'
              examples:
                current_issue:
                  summary: Up-to-date issue
                  value:
                    issueKey: "PROJ-123"
                    projectKey: "PROJ"
                    status: "In Progress"
                    summary: "Implement user authentication"
                    gitFilePath: "PROJ-123.md"
                    jiraUpdatedAt: "2025-09-19T14:10:00Z"
                    syncedAt: "2025-09-19T14:12:00Z"
                    gitCommitHash: "abc123def456"
                    syncStatus: "current"
                stale_issue:
                  summary: Issue needs sync
                  value:
                    issueKey: "PROJ-456"
                    projectKey: "PROJ"
                    status: "Done"
                    summary: "Fix authentication bug"
                    gitFilePath: "PROJ-456.md"
                    jiraUpdatedAt: "2025-09-19T15:00:00Z"
                    syncedAt: "2025-09-19T14:30:00Z"
                    gitCommitHash: "def789ghi012"
                    syncStatus: "needs_sync"
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ProjectSummary:
      type: object
      required:
        - projectKey
        - name
        - status
        - lastSyncTime
        - syncedIssueCount
      properties:
        projectKey:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          description: JIRA project key
          example: "PROJ"
        name:
          type: string
          description: JIRA project name
          example: "Sample Project"
        status:
          type: string
          enum: [Pending, Syncing, Current, Error, Paused]
          description: Current synchronization status
          example: "Current"
        lastSyncTime:
          type: string
          format: date-time
          description: Timestamp of last successful synchronization
          example: "2025-09-19T14:25:00Z"
        syncedIssueCount:
          type: integer
          minimum: 0
          description: Number of issues currently synchronized
          example: 156
        gitRepository:
          type: string
          format: uri
          description: Git repository URL
          example: "git@github.com:company/jira-mirror.git"
    ProjectDetails:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          properties:
            jiraInstance:
              type: object
              description: JIRA instance configuration
              properties:
                baseURL:
                  type: string
                  format: uri
                  description: JIRA instance base URL
                  example: "https://company.atlassian.net"
                proxyEnabled:
                  type: boolean
                  description: Whether HTTP proxy is enabled
                  example: false
            syncConfig:
              type: object
              description: Synchronization configuration
              properties:
                interval:
                  type: string
                  description: Sync polling interval
                  pattern: '^[0-9]+(s|m|h)$'
                  example: "5m"
                activeIssuesOnly:
                  type: boolean
                  description: Whether to sync only active issues
                  example: true
                bootstrap:
                  type: boolean
                  description: Whether bootstrap is enabled
                  example: true
            operandStatus:
              type: object
              description: Status of managed operands
              properties:
                api:
                  $ref: '#/components/schemas/OperandStatus'
                ui:
                  $ref: '#/components/schemas/OperandStatus'
            recentTasks:
              type: array
              description: Recent CDC tasks for this project
              items:
                $ref: '#/components/schemas/TaskSummary'
...
    SyncRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [bootstrap, reconciliation, maintenance]
          description: Type of synchronization operation
          example: "reconciliation"
        forceRefresh:
          type: boolean
          default: false
          description: Whether to ignore last sync timestamps
          example: false
        issueFilter:
          type: string
          description: JQL query to filter issues for sync
          example: "status != Done AND updated >= -7d"
    TaskSummary:
      type: object
      required:
        - id
        - type
        - status
        - projectKey
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        type:
          type: string
          enum: [bootstrap, reconciliation, maintenance]
          description: Type of CDC task
          example: "reconciliation"
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
          description: Current task status
          example: "running"
        projectKey:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          description: JIRA project key
          example: "PROJ"
        createdAt:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2025-09-19T14:15:00Z"
        startedAt:
          type: string
          format: date-time
          description: Task start timestamp
          example: "2025-09-19T14:16:00Z"
        completedAt:
          type: string
          format: date-time
          description: Task completion timestamp
          example: "2025-09-19T14:20:00Z"
        progress:
          $ref: '#/components/schemas/TaskProgress'
...
    TaskResponse:
      type: object
      required:
        - taskId
        - status
        - message
      properties:
        taskId:
          type: string
          format: uuid
          description: Task identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        status:
          type: string
          description: Operation status
          example: "started"
        message:
          type: string
          description: Human-readable status message
          example: "Reconciliation task started successfully"
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error code
          enum: [BadRequest, NotFound, InternalError, ServiceUnavailable, InvalidRequest]
          example: "NotFound"
        message:
          type: string
          description: Human-readable error message
          example: "Project INVALID not found"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-09-19T14:30:00Z"
        details:
          type: object
          description: Additional error context
          additionalProperties: true
          example:
            requestId: "req-123e4567"
            component: "project-service"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for API authentication. In production environments,
        this should be obtained from the cluster's authentication system.
security:
  - bearerAuth: []
...